Create wrapper function to withdraw
